# frontendnew/phonebook-frontend/src/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Jefferson Lab Phonebook",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/groups/{group_id}/members": {
            "get": {
                "tags": [
                    "Working Groups"
                ],
                "summary": "List members of a specific working group",
                "description": "Retrieves a list of members belonging to a specific working group.",
                "operationId": "Working Groups-list_group_members_of_group",
                "parameters": [
                    {
                        "name": "group_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Group Id"
                        }
                    },
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "$ref": "#/components/schemas/GroupRole"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Role"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GroupMemberResponse"
                                    },
                                    "title": "Response Working Groups-List Group Members Of Group"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Working Groups"
                ],
                "summary": "Add a member to a working group",
                "description": "Adds an existing member to a specific working group with a given role.",
                "operationId": "Working Groups-add_member_to_group",
                "parameters": [
                    {
                        "name": "group_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Group Id"
                        }
                    },
                    {
                        "name": "member_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Member Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GroupMemberCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GroupMemberResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GroupMemberCreate": {
                "properties": {
                    "group_id": {
                        "type": "integer",
                        "title": "Group Id"
                    },
                    "member_id": {
                        "type": "integer",
                        "title": "Member Id"
                    },
                    "role": {
                        "$ref": "#/components/schemas/GroupRole"
                    }
                },
                "type": "object",
                "required": [
                    "group_id",
                    "member_id",
                    "role"
                ],
                "title": "GroupMemberCreate"
            },
            "GroupMemberResponse": {
                "properties": {
                    "group_id": {
                        "type": "integer",
                        "title": "Group Id"
                    },
                    "member_id": {
                        "type": "integer",
                        "title": "Member Id"
                    },
                    "role": {
                        "$ref": "#/components/schemas/GroupRole"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "group_id",
                    "member_id",
                    "role",
                    "id"
                ],
                "title": "GroupMemberResponse"
            },
...
            "GroupRole": {
                "type": "string",
                "enum": [
                    "member",
                    "convenor",
                    "co-convenor"
                ],
                "title": "GroupRole"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}